switch to yarn lock, fix sequelize issue
