--- db/models/index.js
@@ -1,5 +1,0 @@
-const db = require('../')
-    , Users = require('./user')
-
-// define associations
-module.exports = db
