+++ package.json
@@ -16,26 +16,25 @@
     "connect-session-sequelize": "^4.1.0",
     "express": "^4.15.3",
     "express-session": "^1.15.3",
-    "has-flag": "^2.0.0",
     "history": "^4.6.3",
     "morgan": "^1.8.2",
     "passport": "^0.3.2",
     "passport-google-oauth": "^1.0.0",
     "passport-local": "^1.0.0",
-    "pg": "^6.2.4",
-    "react": "^15.6.0",
-    "react-dom": "^15.6.0",
+    "pg": "^6.4.1",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
     "react-redux": "^5.0.5",
     "react-router": "^4.1.1",
-    "redux": "^3.7.0",
+    "redux": "^3.7.2",
     "redux-devtools-extension": "^2.13.2",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0",
-    "sequelize": "^4.1.0"
+    "sequelize": "^4.3.2"
   },
   "devDependencies": {
     "babel-core": "^6.25.0",
-    "babel-loader": "^7.0.0",
+    "babel-loader": "^7.1.1",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-register": "^6.24.1",
@@ -43,6 +42,6 @@
     "enzyme": "^2.9.1",
     "mocha": "^3.4.2",
     "supertest": "^3.0.0",
-    "webpack": "^2.6.1"
+    "webpack": "^3.2.0"
   }
 }
