+++ server/api/index.js
@@ -6,8 +6,8 @@ api
   .use('/users', require('./user'))
 
 // define error and send it to error handling middleware in server/index
-api.use((req, res, next) => {
-  const err = new Error('not found')
-  err.status = 404
-  next(err)
-})
+  .use((req, res, next) => {
+    const err = new Error('not found')
+    err.status = 404
+    next(err)
+  })
