+++ server/index.js
@@ -2,7 +2,7 @@ const { join } = require('path')
 		, bodyParser = require('body-parser')
 		, express = require('express')
 		, app = express()
-		, db = require('../db/models')
+		, db = require('../db')
 		, session = require('express-session')
 		, SequelizeStore = require('connect-session-sequelize')(session.Store)
 		, dbStore = new SequelizeStore({db})
